Introduce a way to add protoc options to protobuf definitions
The MetadataProcessor takes a model and extract a per namespace
map of options. Before rendering a namespace we retrieve the options
for a namespace and add them to the CompilationUnit.

Encoded as follows:

1. we use smithy metadata proto_options
2. we use an array of objects so that multiple proto_options can be concatenated
3. the options are defined per namespace
4. the options are String -> String

Regarding 4), I chose this encoding over String -> Any(bool, string, float, integer, identifier) because of the identifier.

You can do `option thing = "that";` but you can also do:
`option thing = THAT;` where THAT would be a constant that protoc knows
about.

To workaround the issue, the value is always a string, so if you want
the former, you do `"thing": "\"that\""`, and if you want the latter
you do: `"thing": "THAT"`. As such, if you want a `bool` like `true`, you
do `"thing": "true"`.

I'm open to revisit this if you have another proposition.