Fix enum support
In this commit, we diverge from the specification. I've commented
on a similar issue that was resolved a few days ago. My hopes
are that a similar resolution is applied and the divergence is no
more.

In short, the grammar states that a value assignment should look
like this:

`ValueAssignment = *SP "=" *SP NodeValue BR`

But this BR is counter-intuitive and probably what's intended
is a `*WS` which is what this commits implements.

This fixes formatter issues with Enum and Structure which both
can define ValueAssignment on their members