Render top-level union
I spent some time thinking about it and I think this is a
better solution to what we had in place. Previously, for unions,
we would use oneOf and force user to use the Union in only one
place. If used in multiple places, we would throw and if unused,
we would just not render it.

This approach does the following:

1. render every union as a top level message with one member - the oneOf, named
`definition`
2. all other shapes pointing to a union points to this top level
message instead