Use toSet(clazz) to find shapes
When using `model.getShape` with a shapeId and relying on
whether or the shape is there, we risk not finding shapes
because of a bad shape id.

In this case, smithy.api#Timestamp does not exist, so
the timestamp shape was never _used_ according to this logic.
This resulted in the output protobuf models not having
the smithytranslate.timestamp message exported