Allow empty enum
The IDL says: `"{" [WS] 1*(EnumShapeMember [WS]) "}"` which means
atleast one member is required. In this case, I've allowed the
implementation to drift from the spec because I don't think
the parser should be validating business rules like that. From
our point of view, it's hard to tell if the enum is empty when
mixins could be adding members or not.